Split up Showtimes
Little bit of renaming as part of the refactor.
