+++ src/arx7.js
@@ -1,13 +1,14 @@
 import debug from 'debug';
 import pkg from '../package';
 
-import { Blame } from './commands/blame.js';
 import { Choose } from './commands/choose.js';
 import { Imgur } from './commands/imgur.js';
 import { Order } from './commands/order.js';
 import { Reply } from './commands/reply.js';
 import { Same } from './commands/same.js';
-import { Showtimes } from './commands/showtimes.js';
+import { ShowtimesBlame } from './commands/showtimes_blame.js';
+import { ShowtimesRelease } from './commands/showtimes_release.js';
+import { ShowtimesStatus } from './commands/showtimes_status.js';
 import { Time } from './commands/time.js';
 import { Twitter } from './commands/twitter.js';
 import { Youtube } from './commands/youtube.js';
@@ -20,13 +21,14 @@ const BAN_RETRY_DELAY = 1000 * 60 * 3;
 export class ARX7 {
   constructor(client, config) {
     this.commands = [
-      new Blame(client),
       new Choose(client),
       new Imgur(client),
       new Order(client),
       new Reply(client),
       new Same(client),
-      new Showtimes(client, this.isAuthorized),
+      new ShowtimesBlame(client),
+      new ShowtimesRelease(client, this.isAuthorized),
+      new ShowtimesStatus(client, this.isAuthorized),
       new Time(client),
       new Twitter(client),
       new Youtube(client),
